Fix regex, improve error messages
Running autotest to validate firmware upgrade
uncovered a cut&paste error in the the firmware regex.
The error prevents firmware upgrade.

Also when the package is missing a rule file, no error
was reported.

BUG=chromium:328624
TEST=unit test
TEST=apply firmware packages to upgrade/downgrade and invalid package

Reviewed-on: https://chromium-review.googlesource.com/186753
Commit-Queue: Gwendal Grignou <gwendal@chromium.org>
Tested-by: Gwendal Grignou <gwendal@chromium.org>

+++ disk_updater/scripts/chromeos-disk-firmware-update.sh
@@ -178,7 +178,7 @@ disk_info() {
     | sed -re 's/ +$//' -e 's/[ -]/_/g')
   disk_fw_rev=$(sed -nre \
       '/^\t+Firmware/s|\t+Firmware Revision: +(.*)|\1|p' "${hdparm_out}" \
-    | sed -re 's/ \+$//' -e 's/[ -]/_/g')
+    | sed -re 's/ +$//' -e 's/[ -]/_/g')
   if [ -z "${disk_model}" -o -z "${disk_fw_rev}" ]; then
     return 1
   fi
@@ -231,10 +231,6 @@ disk_upgrade_devices() {
   local rc=0
   local tries=0
 
-  if [ ! -f "${disk_rules}" ]; then
-    log_msg "Unable to find rules file in ${FLAGS_fw_package_dir}"
-    return 120
-  fi
   shift # skip disk rules parameters.
   for device in "$@"; do
     sucess=""
@@ -322,7 +318,11 @@ main() {
     erase_tmp_dir=${FLAGS_TRUE}
     FLAGS_tmp_dir=$(mktemp -d)
   fi
-  disk_rules=${FLAGS_tmp_dir}/disk_rules
+  if [ ! -f "${disk_rules_raw}" ]; then
+    log_msg "Unable to find rules file in ${FLAGS_fw_package_dir}"
+    return 120
+  fi
+  disk_rules=${FLAGS_tmp_dir}/rules
 
   # remove unnecessary lines
   sed '/^#/d;/^[[:space:]]*$/d' "${disk_rules_raw}" > "${disk_rules}"
